rename some files, adjust more fonts
