+++ src/App.js
@@ -3,7 +3,7 @@ import { ThemeProvider } from "@mui/material/styles";
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from "./pages/home/Home.jsx";
 import About from "./pages/about/About.jsx";
-import NavigationBar from "./components/NavigationBar";
+import NavigationBar from "./components/navigationBar/NavigationBar.jsx";
 import Footer from "./components/Footer";
 import theme from "./theme/theme.js";
