+++ src/components/Footer.jsx
@@ -1,9 +1,12 @@
 import * as React from "react";
+import Typography from "@mui/material/Typography";
 
 const Footer = () => {
   return (
     <footer>
-      <p>© {new Date().getFullYear()} Antonio Franco</p>
+      <Typography variant="body2">
+      © {new Date().getFullYear()} Antonio Franco
+      </Typography>
     </footer>
   );
 };
