+++ src/components/aboutWindow/aboutWindow.css
