+++ src/components/navigationBar/NavigationBar.jsx
@@ -2,18 +2,15 @@ import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
 import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
-import Button from "@mui/material/Button";
 import IconButton from "@mui/material/IconButton";
 import MenuIcon from "@mui/icons-material/Menu";
 import Drawer from '@mui/material/Drawer';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
-import Hidden from '@mui/material/Hidden';
 import { useMediaQuery } from '@mui/material';
+import "./navigationBar.css";
 
 
 const NavigationBar = () => {
@@ -30,7 +27,6 @@ const NavigationBar = () => {
     { text: 'About', link: '/about' },
     { text: 'Projects', link: '/projects' },
     { text: 'Contact', link: '/contact' },
-    // Add more menu items as needed
   ];
 
   const drawer = (
@@ -44,7 +40,7 @@ const NavigationBar = () => {
   );
 
   return (
-    <div>
+    <div className="navigation-bar">
       <AppBar position="fixed">
         <Toolbar className="nav-bar">
           {isMobile && (
@@ -81,26 +77,6 @@ const NavigationBar = () => {
         )}
       </nav>
     </div>
-    // <Box sx={{ flexGrow: 1 }}>
-    //   <AppBar position="static">
-    //     <Toolbar>
-    //       <IconButton
-    //         size="large"
-    //         edge="start"
-    //         color="inherit"
-    //         aria-label="menu"
-    //         sx={{ mr: 2 }}
-    //       >
-    //         <MenuIcon />
-    //       </IconButton>
-    //       <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-    //         Antonio Franco
-    //       </Typography>
-    //       <Button color="inherit">Home</Button>
-    //       <Button color="inherit">Running</Button>
-    //     </Toolbar>
-    //   </AppBar>
-    // </Box>
   );
 };
