+++ src/components/navigationBar/navigationBar.css
@@ -1,0 +1,3 @@
+.navigation-bar span {
+    color: #ffffff;
+}
