+++ src/pages/about/about.css
@@ -3,5 +3,5 @@
     flex-direction: row;
     justify-content: center;
     align-items: center;
-    margin: 40px;
+    margin: 80px;
 }
